!: don't create a subshell when spawning new processes
This change fixes the support for hanging-process-killer-program.
Before this change, hanging-process-killer-program would be given the
pid of the parent shell of salmonella.  Thus,
hanging-process-killer-program would monitor the shell process, which
in turn would fork salmonella. hanging-process-killer-program would
end up killing salmonella for considering that it was hanging (usually
salmonella's execution exceeds the timeout configured for
hanging-process-killer-program).  The right thing to do is to spawn
processes without an intermediate shell -- that's what this commit
does.  With this, hanging-process-killer-program will monitor the
children of the actual salmonella process.