Add pre-built-chicken parameter
If this parameter is not #f, salmonella-run-publish will use the
CHICKEN installation pointed by the value of this parameter instead of
building CHICKEN from git.  CHICKEN build-related hooks will be just
executed in order, without any intermediate steps between them.

This parameter is mostly useful to speed-up tests of salmonella and/or
salmonella-run-publish itself.